refactor(system): system 패키지 생성, 시스템과 관련된 클래스 분류
