+++ src/main/java/baseball/system/ExceptionMessage.java
@@ -1,4 +1,4 @@
-package baseball.utils;
+package baseball.system;
 
 public class ExceptionMessage {
     public static final String NOT_DISTINCT = "중복된 자리수가 존재합니다";
