+++ src/main/java/baseball/verifier/NumberVerifier.java
@@ -1,7 +1,7 @@
 package baseball.verifier;
 
-import baseball.utils.ExceptionMessage;
-import baseball.utils.SystemConstant;
+import baseball.system.ExceptionMessage;
+import baseball.system.SystemConstant;
 
 public class NumberVerifier implements Verifier {
     @Override
